Remove tracked sphinx files
